Add files via upload
It is a food delivery system which is used to find the shortest distance for delivering food at different delivery locations which are considered as vertices . We are using two algorithms Floyd warshall and Nearest neighbour approach as if number of vertices (deliveries) are more than 5 than Floyd will be used and if number of deliveries are greater than 5 than nearest neighbour approach will be used. Also we can generate bill , in this crono header file is used to print a time stamp on bill receit. Quicksort is used to sort the delivery boys depending on the number of deliveries they have made . Also I have changed console colour and added other facilities to make this program user friendly .